+++ packages/sqflite/example/lib/type_test_page.dart
@@ -1,8 +1,8 @@
-import 'dart:async';
 import 'dart:math';
 
 import 'package:flutter/foundation.dart';
 import 'package:sqflite/sqflite.dart';
+import 'package:sqflite/utils/utils.dart';
 import 'package:sqflite_tizen_example/utils.dart';
 
 import 'test_page.dart';
@@ -41,23 +41,41 @@ class TypeTestPage extends TestPage {
       //devPrint('2^33: ${await getValue(id)}');
       expect(await getValue(id), pow(2, 33));
 
-      id = await insertValue(pow(2, 62));
-      //devPrint('2^62: ${pow(2, 62)} ${await getValue(id)}');
-      expect(
-        await getValue(id),
-        pow(2, 62),
-        reason: '2^62: ${pow(2, 62)} ${await getValue(id)}',
-      );
+      // about web limits
+      id = await insertValue(pow(2, 53));
+      expect(await getValue(id), pow(2, 53));
 
-      var value = pow(2, 63).round() - 1;
-      id = await insertValue(value);
-      //devPrint('${value} ${await getValue(id)}');
-      expect(await getValue(id), value, reason: '$value ${await getValue(id)}');
+      if (!kIsWeb) {
+        // Not for web
+        id = await insertValue(pow(2, 62));
+        //devPrint('2^62: ${pow(2, 62)} ${await getValue(id)}');
+        expect(
+          await getValue(id),
+          pow(2, 62),
+          reason: '2^62: ${pow(2, 62)} ${await getValue(id)}',
+        );
 
-      value = -(pow(2, 63)).round();
-      id = await insertValue(value);
-      //devPrint('${value} ${await getValue(id)}');
-      expect(await getValue(id), value, reason: '$value ${await getValue(id)}');
+        var value = pow(2, 63).round() - 1;
+        id = await insertValue(value);
+        //devPrint('${value} ${await getValue(id)}');
+        expect(
+          await getValue(id),
+          value,
+          reason: '$value ${await getValue(id)}',
+        );
+
+        value = -(pow(2, 63)).round();
+        id = await insertValue(value);
+        //devPrint('${value} ${await getValue(id)}');
+        expect(
+          await getValue(id),
+          value,
+          reason: '$value ${await getValue(id)}',
+        );
+      }
+      var bigValue = 1234567890123456;
+      id = await insertValue(bigValue);
+      expect(await getValue(id), bigValue);
       /*
       id = await insertValue(pow(2, 63));
       devPrint('2^63: ${pow(2, 63)} ${await getValue(id)}');
@@ -91,12 +109,18 @@ class TypeTestPage extends TestPage {
       // big float
       id = await insertValue(1 / 3);
       expect(await getValue(id), 1 / 3);
-      id = await insertValue(pow(2, 63) + .1);
-      expect(await getValue(id), pow(2, 63) + 0.1);
+      if (!kIsWeb) {
+        id = await insertValue(pow(2, 63) + .1);
+        expect(await getValue(id), pow(2, 63) + 0.1);
 
-      // integer?
-      id = await insertValue(pow(2, 62));
-      expect(await getValue(id), pow(2, 62));
+        // integer?
+        id = await insertValue(pow(2, 62));
+        expect(await getValue(id), pow(2, 62));
+      }
+      var bigValue = 1234567890123456789.0;
+      id = await insertValue(bigValue);
+      expect(await getValue(id), bigValue);
+
       await data.db.close();
     });
 
@@ -128,7 +152,7 @@ class TypeTestPage extends TestPage {
     });
 
     test('blob', () async {
-      // await Sqflite.devSetDebugModeOn(true);
+      // databaseFactory = databaseFactory.debugQuickLoggerWrapper();
       final path = await initDeleteDb('type_blob.db');
       data.db = await openDatabase(
         path,
@@ -189,18 +213,17 @@ class TypeTestPage extends TestPage {
         // try blob lookup using hex
         rows = await data.db.rawQuery(
           'SELECT * FROM Test WHERE hex(value) = ?',
-          [Sqflite.hex(blob1234)],
+          [hex(blob1234)],
         );
         expect(rows.length, 1);
         expect(rows[0]['id'], 3);
 
         // Insert empty blob
-        // TODO: Fix this behavior
-        // final blobEmpty = Uint8List(0);
-        // id = await insertValue(blobEmpty);
-        // value = await getValue(id);
-        // expect(value, const TypeMatcher<Uint8List>());
-        // expect(value, isEmpty);
+        final blobEmpty = Uint8List(0);
+        id = await insertValue(blobEmpty);
+        value = await getValue(id);
+        expect(value, const TypeMatcher<Uint8List>());
+        expect(value, isEmpty);
       } finally {
         await data.db.close();
       }
@@ -251,6 +274,9 @@ class TypeTestPage extends TestPage {
           await insertValue(DateTime.fromMillisecondsSinceEpoch(1234567890));
         } on ArgumentError catch (_) {
           failed = true;
+        } on UnsupportedError catch (_) {
+          // this happens on the web
+          failed = true;
         }
         expect(failed, true);
       } finally {
@@ -274,6 +300,9 @@ class TypeTestPage extends TestPage {
         try {
           await insertValue(true);
         } on ArgumentError catch (_) {
+          failed = true;
+        } on DatabaseException catch (_) {
+          // this happens on the web
           failed = true;
         }
         if (supportsCompatMode) {